requirements.txt
langgraph
langchain-openai
langchain-community
duckduckgo-search
python-dotenv
pydantic